Fix ut

+++ dubbo-common/src/main/java/org/apache/dubbo/rpc/model/ModuleModel.java
@@ -47,35 +47,37 @@ public class ModuleModel extends ScopeModel {
 
     @Override
     public void destroy() {
-        List<ConsumerModel> consumerModels = serviceRepository.getReferredServices();
+        if (serviceRepository != null) {
+            List<ConsumerModel> consumerModels = serviceRepository.getReferredServices();
 
-        for (ConsumerModel consumerModel : consumerModels) {
-            try {
-                if (consumerModel.getReferenceConfig() != null) {
-                    consumerModel.getReferenceConfig().destroy();
-                } else if (consumerModel.getDestroyCaller() != null) {
-                    consumerModel.getDestroyCaller().call();
+            for (ConsumerModel consumerModel : consumerModels) {
+                try {
+                    if (consumerModel.getReferenceConfig() != null) {
+                        consumerModel.getReferenceConfig().destroy();
+                    } else if (consumerModel.getDestroyCaller() != null) {
+                        consumerModel.getDestroyCaller().call();
+                    }
+                } catch (Throwable t) {
+                    logger.error("Unable to destroy consumerModel.", t);
                 }
-            } catch (Throwable t) {
-                logger.error("Unable to destroy consumerModel.", t);
             }
-        }
 
-        List<ProviderModel> exportedServices = serviceRepository.getExportedServices();
-        for (ProviderModel providerModel : exportedServices) {
-            try {
-                if (providerModel.getServiceConfig() != null) {
-                    providerModel.getServiceConfig().unexport();
-                } else if (providerModel.getDestroyCaller() != null) {
-                    providerModel.getDestroyCaller().call();
+            List<ProviderModel> exportedServices = serviceRepository.getExportedServices();
+            for (ProviderModel providerModel : exportedServices) {
+                try {
+                    if (providerModel.getServiceConfig() != null) {
+                        providerModel.getServiceConfig().unexport();
+                    } else if (providerModel.getDestroyCaller() != null) {
+                        providerModel.getDestroyCaller().call();
+                    }
+                } catch (Throwable t) {
+                    logger.error("Unable to destroy providerModel.", t);
                 }
-            } catch (Throwable t) {
-                logger.error("Unable to destroy providerModel.", t);
             }
-        }
 
-        serviceRepository.destroy();
-        serviceRepository = null;
+            serviceRepository.destroy();
+            serviceRepository = null;
+        }
 
         super.destroy();
         // TODO destroy module resources
